Admin lists pending extensions.

+++ src/main/java/com/telerik/extension_repository/models/bindingModels/extensions/RelatedExtensionModel.java
@@ -1,13 +1,13 @@
 package com.telerik.extension_repository.models.bindingModels.extensions;
 
 public class RelatedExtensionModel {
-    private Integer id;
+    private Long id;
 
-    public Integer getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
